styling button send email
