+++ fe-generate-invoice/public/index.html
@@ -10,7 +10,7 @@
     <link rel="icon" href="/logo-ginap.png" />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css" />
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-iYQeCzEYFbKjA/T2uDLTpkwGzCiq6soy8tYaI1GyVh/UjpbCx/TYkiZhlZB6+fzT" crossorigin="anonymous" />
-    <title>React App</title>
+    <title>Ginap</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
